wip: move event listener into SharedUtils
